webpki: move WebPkiServerVerifier to server_verifier.rs
This commit moves the existing `WebPkiServerVerifier` type out of
`verify.rs` and into `server_verifier.rs`. This matches the new
`client_verifier.rs` module introduced in a previous commit and paves
the way for having a single module location for both the verifier type
and a builder for making it.
